+++ auth/proofOfPossession_test.go
@@ -4,9 +4,10 @@
 package auth
 
 import (
+	"testing"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 var derCertGenerated = generateDERCert()
