+++ auth/token_test.go
@@ -8,9 +8,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/google/uuid"
 	"github.com/lestrrat-go/jwx/jwt"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestToken_getClaimAsString(t *testing.T) {
@@ -197,13 +195,4 @@ func TestOIDCClaims_getSAPIssuer(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestNewTokenFromClaims(t *testing.T) {
-	userUUID := uuid.NewString()
-	m := map[string]interface{}{"user_uuid": userUUID}
-	claims, err := NewTokenFromClaims(m)
-	assert.NoError(t, err)
-
-	assert.Equal(t, userUUID, claims.UserUUID(), "UserUUID() got = %v, want %v", claims.UserUUID(), userUUID)
 }
